Add public network to Vagrant instances.
Each instance is created with 2 private networks today.
We want to add another network for data, and potentially another for
access later on.

This solution will add an if..else clause for the Vagrantfile to include
public network with the specified parameters in the topology.yml

topology.yaml: added lightbits network interfaces which is taken as
public_network with all the relevant information.